Initial submit
So far all this does is navigate through a directory and list the names of all of the HTML files (or at least ones that probably are). Now that this works, I will be moving towards parsing all of those files to find the instance of "<form>" Once I do, I will gather all of the input variables and their values, then input them into a dictionary.