Added variable find
This version finds all of the html files from the folder created by wget. It them scans all of those files for instances of "<input>" sections. It then looks for all instances of either "id" or "name" in order to gather all of the variable names. Then it looks for instances of "value." When it finds each of those, it puts them into a text file, which will be the dictionary used by configfileio to update form fields. I still need to clean this up and get rid of any excess code.