Attempts to make this even more generic
This version is a very stripped away webserver. As of now it does not work for the Koyo if you attempt to submit data. It breaks and does not allow you to visit any more pages. It also does not work at all for the Allen-Bradley (I would have thought it would at least bring up the main page but it doesn't.