Works with the working version of configfileio
This version is a bit ugly with all of the print statements. But it works with the current version of configfileio. This version still will not display certain pages, despite the templates being formatted properly. I need to see if it goes beyond templates to something more complicated, like the format of some of the inputs (maybe a time to use regex?)